+++ RevenueCat/Scripts/Purchases.cs
@@ -73,141 +73,6 @@ public class Purchases : MonoBehaviour
         public abstract void PurchaserInfoReceived(PurchaserInfo purchaserInfo);
     }
 
-    private class PurchasesWrapperNoop : IPurchasesWrapper
-    {
-        public void Setup(string gameObject, string apiKey, string appUserId, bool observerMode)
-        {
-
-        }
-
-        public void AddAttributionData(int network, string data, string networkUserId)
-        {
-
-        }
-
-        public void GetProducts(string[] productIdentifiers, string type = "subs")
-        {
-
-        }
-
-        public void MakePurchase(string productIdentifier, string type = "subs", string oldSku = null)
-        {
-
-        }
-
-        public void PurchaseProduct(string productIdentifier, string type = "subs", string oldSku = null, ProrationMode prorationMode = ProrationMode.UnknownSubscriptionUpgradeDowngradePolicy)
-        {
-
-        }
-
-        public void PurchasePackage(Purchases.Package packageToPurchase, string oldSku = null, ProrationMode prorationMode = ProrationMode.UnknownSubscriptionUpgradeDowngradePolicy)
-        {
-
-        }
-
-        public void RestoreTransactions()
-        {
-
-        }
-
-        public void CreateAlias(string newAppUserId)
-        {
-
-        }
-
-        public void Identify(string appUserId)
-        {
-
-        }
-
-        public void Reset()
-        {
-
-        }
-
-        public void SetFinishTransactions(bool finishTransactions)
-        {
-
-        }
-
-        public void SetAllowSharingStoreAccount(bool allow)
-        {
-
-        }
-
-        public string GetAppUserId()
-        {
-            return null;
-        }
-
-        public void SetDebugLogsEnabled(bool enabled)
-        {
-
-        }
-
-        public void GetPurchaserInfo()
-        {
-
-        }
-
-        public void GetOfferings()
-        {
-
-        }
-
-        public void SyncPurchases()
-        {
-
-        }
-
-        public void SetAutomaticAppleSearchAdsAttributionCollection(bool enabled)
-        {
-
-        }
-
-
-        public bool IsAnonymous()
-        {
-            return false;
-        }
-
-        public void CheckTrialOrIntroductoryPriceEligibility(string[] productIdentifiers)
-        {
-
-        }
-
-        public void InvalidatePurchaserInfoCache()
-        {
-
-        }
-
-        public void SetAttributes(Dictionary<string, string> attributes)
-        {
-
-        }
-
-        public void SetEmail(string email)
-        {
-
-        }
-
-        public void SetPhoneNumber(string phoneNumber)
-        {
-
-        }
-
-        public void SetDisplayName(string displayName)
-        {
-
-        }
-
-        public void SetPushToken(string token)
-        {
-
-        }
-
-    }
-
     /*
      * PurchaserInfo encapsulate the current status of subscriber. 
      * Use it to determine which entitlements to unlock, typically by checking
